Address maintainer review feedback
- Move build_dateandtime and parse_dateandtime to utils module
- Merge datetime tests into test_init.py (match code file convention)
- Use pytest fixture for fake _request_args setup

Made-with: Cursor
