+++ brother/utils.py
@@ -1,9 +1,12 @@
 """Utils for Brother."""
 
 import asyncio
+from datetime import datetime
 
 from pysnmp.hlapi.v3arch.asyncio import SnmpEngine
 from pysnmp.hlapi.varbinds import MibViewControllerManager
+
+from .const import DATEANDTIME_MIN_LENGTH
 
 
 async def async_get_snmp_engine() -> SnmpEngine:
@@ -29,3 +32,28 @@ def bytes_to_hex_string(data: bytes) -> str:
     # More efficient than join with list comprehension
     # Remove last 2 characters (last byte in hex) for checksum
     return data[:-1].hex()
+
+
+def build_dateandtime(dt: datetime) -> bytes:
+    """Encode a datetime as an 8-byte SNMP DateAndTime value (RFC 2579)."""
+    return (
+        dt.year.to_bytes(2, "big")
+        + bytes([dt.month, dt.day, dt.hour, dt.minute, dt.second, 0])
+    )
+
+
+def parse_dateandtime(raw: bytes) -> datetime | None:
+    """Decode an SNMP DateAndTime value into a naive datetime.
+
+    The 8-byte DateAndTime format carries no timezone offset, so the
+    returned datetime is naive and represents the printer's local clock.
+    """
+    if len(raw) < DATEANDTIME_MIN_LENGTH:
+        return None
+    year = int.from_bytes(raw[0:2], "big")
+    try:
+        return datetime(  # noqa: DTZ001
+            year, raw[2], raw[3], raw[4], raw[5], raw[6]
+        )
+    except ValueError:
+        return None
